requirements.txt
pandas==2.3.3
fastapi==0.121.1
pydantic==2.12.4
mysql-connector-python==9.5.0
streamlit==1.51.0
requests==2.32.5
uvicorn==0.38.0
pytest==9.0.0